# Optimize LIKE for more relaxed patterns