--- a/velox/benchmarks/basic/LikeBenchmark.cpp
+++ b/velox/benchmarks/basic/LikeBenchmark.cpp
@@ -35,59 +35,87 @@ int main(int argc, char** argv) {
   // Register the scalar functions.
   prestosql::registerAllScalarFunctions("");
 
-  // exec::register
   ExpressionBenchmarkBuilder benchmarkBuilder;
   const vector_size_t vectorSize = 1000;
   auto vectorMaker = benchmarkBuilder.vectorMaker();
 
-  auto makeInput =
-      [&](vector_size_t vectorSize, bool padAtHead, bool padAtTail) {
-        return vectorMaker.flatVector<std::string>(vectorSize, [&](auto row) {
-          // Strings in even rows contain/start with/end with a_b_c depends on
-          // value of padAtHead && padAtTail.
-          if (row % 2 == 0) {
-            auto padding = std::string(row / 2 + 1, 'x');
-            if (padAtHead && padAtTail) {
-              return fmt::format("{}a_b_c{}", padding, padding);
-            } else if (padAtHead) {
-              return fmt::format("{}a_b_c", padding);
-            } else if (padAtTail) {
-              return fmt::format("a_b_c{}", padding);
-            } else {
-              return std::string("a_b_c");
-            }
-          } else {
-            return std::string(row, 'x');
-          }
-        });
-      };
+  auto makeInput = [&](vector_size_t vectorSize,
+                       bool padAtHead,
+                       bool padAtTail,
+                       std::string content = "a_b_c",
+                       std::string paddingStr = "xxx") {
+    return vectorMaker.flatVector<std::string>(vectorSize, [&](auto row) {
+      // Strings in even rows contain/start with/end with a_b_c depends on
+      // value of padAtHead && padAtTail.
+
+      // Calculates the padding.
+      std::ostringstream os;
+      for (auto i = 0; i < row / 2 + 1; ++i) {
+        os << paddingStr;
+      }
+      auto padding = os.str();
+
+      if (row % 2 == 0) {
+        if (padAtHead && padAtTail) {
+          return fmt::format("{}{}{}", padding, content, padding);
+        } else if (padAtHead) {
+          return fmt::format("{}{}", padding, content);
+        } else if (padAtTail) {
+          return fmt::format("{}{}", content, padding);
+        } else {
+          return content;
+        }
+      } else {
+        // Yes, two padding concatenated, since we have a '/2' above.
+        return padding + padding;
+      }
+    });
+  };
 
   auto substringInput = makeInput(vectorSize, true, true);
   auto prefixInput = makeInput(vectorSize, false, true);
+  auto prefixUnicodeInput = makeInput(vectorSize, false, true, "你_好_啊");
   auto suffixInput = makeInput(vectorSize, true, false);
+  auto suffixUnicodeInput = makeInput(vectorSize, true, false, "你_好_啊");
 
   benchmarkBuilder
       .addBenchmarkSet(
-          "like_substring", vectorMaker.rowVector({"col0"}, {substringInput}))
-      .addExpression("like_substring", R"(like(col0, '%a\_b\_c%', '\'))")
+          "substring", vectorMaker.rowVector({"col0"}, {substringInput}))
+      .addExpression("substring", R"(like(col0, '%a\_b\_c%', '\'))")
       .addExpression("strpos", R"(strpos(col0, 'a_b_c') > 0)");
 
   benchmarkBuilder
       .addBenchmarkSet(
-          "like_prefix", vectorMaker.rowVector({"col0"}, {prefixInput}))
-      .addExpression("like_prefix", R"(like(col0, 'a\_b\_c%', '\'))")
+          "prefix",
+          vectorMaker.rowVector(
+              {"col0", "col1"}, {prefixInput, prefixUnicodeInput}))
+      .addExpression("prefix", R"(like(col0, 'a\_b\_c%', '\'))")
+      .addExpression("relaxed_prefix_1", R"(like(col0, 'a\__\_c%', '\'))")
+      .addExpression("relaxed_prefix_2", R"(like(col0, '_\__\_c%', '\'))")
+      .addExpression(
+          "relaxed_prefix_unicode_1", R"(like(col1, '你\__\_啊%', '\'))")
+      .addExpression(
+          "relaxed_prefix_unicode_2", R"(like(col1, '_\__\_啊%', '\'))")
       .addExpression("starts_with", R"(starts_with(col0, 'a_b_c'))");
 
   benchmarkBuilder
       .addBenchmarkSet(
-          "like_suffix", vectorMaker.rowVector({"col0"}, {suffixInput}))
-      .addExpression("like_suffix", R"(like(col0, '%a\_b\_c', '\'))")
+          "suffix",
+          vectorMaker.rowVector(
+              {"col0", "col1"}, {suffixInput, suffixUnicodeInput}))
+      .addExpression("suffix", R"(like(col0, '%a\_b\_c', '\'))")
+      .addExpression("relaxed_suffix_1", R"(like(col0, '%a\__\_c', '\'))")
+      .addExpression("relaxed_suffix_2", R"(like(col0, '%_\__\_c', '\'))")
+      .addExpression(
+          "relaxed_suffix_unicode_1", R"(like(col1, '%你\__\_啊', '\'))")
+      .addExpression(
+          "relaxed_suffix_unicode_2", R"(like(col1, '%_\__\_啊', '\'))")
       .addExpression("ends_with", R"(ends_with(col0, 'a_b_c'))");
 
   benchmarkBuilder
       .addBenchmarkSet(
-          "like_generic", vectorMaker.rowVector({"col0"}, {substringInput}))
-      .addExpression("like_generic", R"(like(col0, '%a%b%c'))");
+          "generic", vectorMaker.rowVector({"col0"}, {substringInput}))
+      .addExpression("generic", R"(like(col0, '%a%b%c'))");
 
   benchmarkBuilder.registerBenchmarks();
   benchmarkBuilder.testBenchmarks();