@xumingming 2024-01-30
cc @mbasmanova 

The difference between this and previous PR is the third commit([Fix kRelaxedFixed pattern](https://github.com/facebookincubator/velox/pull/8594/commits/ffe4f6b2ba99fdb371ee9768a5ed46e9f23a0389)). 

`matchRelaxedFixedForward` is a help function which returns a `std::pair`
- first: a bool indicates whether matches the pattern.
- second: an integer indicates where is cursor in the input when we finished matching if 'first' is true, -1 otherwise.

So for `kRelaxedPrefix` we just use `matchRelaxedFixedForward` is enough. For`kRelaxedFixed` we additionally check `matchResult.second == input.size()`.